added v2 of protocol, default for C++. need to update v1 before release

+++ tools/general-loader/utf8-like-int-codec.c
@@ -1,4 +1,28 @@
-/* utf8-like-int-codec.c */
+/*===========================================================================
+*
+*                            PUBLIC DOMAIN NOTICE
+*               National Center for Biotechnology Information
+*
+*  This software/database is a "United States Government Work" under the
+*  terms of the United States Copyright Act.  It was written as part of
+*  the author's official duties as a United States Government employee and
+*  thus cannot be copyrighted.  This software/database is freely available
+*  to the public for use. The National Library of Medicine and the U.S.
+*  Government have not placed any restriction on its use or reproduction.
+*
+*  Although all reasonable efforts have been taken to ensure the accuracy
+*  and reliability of the software and data, the NLM and the U.S.
+*  Government do not and cannot warrant the performance or results that
+*  may be obtained by using this software or data. The NLM and the U.S.
+*  Government disclaim all warranties, express or implied, including
+*  warranties of performance, merchantability or fitness for any particular
+*  purpose.
+*
+*  Please cite the author in any work or product based on this material.
+*
+* ===========================================================================
+*
+*/
 
 #include "utf8-like-int-codec.h"
 
@@ -13,10 +37,10 @@
 #define MAX_VALUE_BYTE_7 0xFFFFFFFFF
 
 
-int64_t encode_uint16 ( uint16_t value_to_encode, uint8_t* buf_start, uint8_t* buf_xend )
+int encode_uint16 ( uint16_t value_to_encode, uint8_t* buf_start, uint8_t* buf_xend )
 {
     size_t dst_buf_size = buf_xend - buf_start;
-    int64_t ret = CODEC_UNKNOWN_ERROR;
+    int ret = CODEC_UNKNOWN_ERROR;
 
     /* optimization: process 1-byte case in the very beginning */
     if (value_to_encode <= MAX_VALUE_BYTE_1)
@@ -53,10 +77,10 @@ int64_t encode_uint16 ( uint16_t value_to_encode, uint8_t* buf_start, uint8_t* b
     return ret;
 }
 
-int64_t decode_uint16 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint16_t* ret_decoded )
+int decode_uint16 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint16_t* ret_decoded )
 {
     size_t src_buf_size = buf_xend - buf_start;
-    int64_t ret = CODEC_INVALID_FORMAT;
+    int ret = CODEC_INVALID_FORMAT;
 
     if (src_buf_size < 1)
         return CODEC_INSUFFICIENT_BUFFER;
@@ -119,10 +143,10 @@ int64_t decode_uint16 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint1
 }
 
 
-int64_t encode_uint32 ( uint32_t value_to_encode, uint8_t* buf_start, uint8_t* buf_xend )
+int encode_uint32 ( uint32_t value_to_encode, uint8_t* buf_start, uint8_t* buf_xend )
 {
     size_t dst_buf_size = buf_xend - buf_start;
-    int64_t ret = CODEC_UNKNOWN_ERROR;
+    int ret = CODEC_UNKNOWN_ERROR;
 
     if (value_to_encode <= MAX_VALUE_BYTE_1)
     {
@@ -182,10 +206,10 @@ int64_t encode_uint32 ( uint32_t value_to_encode, uint8_t* buf_start, uint8_t* b
     return ret;
 }
 
-int64_t decode_uint32 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint32_t* ret_decoded )
+int decode_uint32 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint32_t* ret_decoded )
 {
     size_t src_buf_size = buf_xend - buf_start;
-    int64_t ret = CODEC_INVALID_FORMAT;
+    int ret = CODEC_INVALID_FORMAT;
 
     if (src_buf_size < 1)
         return CODEC_INSUFFICIENT_BUFFER;
@@ -307,10 +331,10 @@ int64_t decode_uint32 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint3
 }
 
 
-int64_t encode_uint64 ( uint64_t value_to_encode, uint8_t* buf_start, uint8_t* buf_xend )
+int encode_uint64 ( uint64_t value_to_encode, uint8_t* buf_start, uint8_t* buf_xend )
 {
     size_t dst_buf_size = buf_xend - buf_start;
-    int64_t ret = CODEC_UNKNOWN_ERROR;
+    int ret = CODEC_UNKNOWN_ERROR;
 
     if (value_to_encode <= MAX_VALUE_BYTE_1)
     {
@@ -416,10 +440,10 @@ int64_t encode_uint64 ( uint64_t value_to_encode, uint8_t* buf_start, uint8_t* b
     return ret;
 }
 
-int64_t decode_uint64 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint64_t* ret_decoded )
+int decode_uint64 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint64_t* ret_decoded )
 {
     size_t src_buf_size = buf_xend - buf_start;
-    int64_t ret = CODEC_INVALID_FORMAT;
+    int ret = CODEC_INVALID_FORMAT;
 
     if (src_buf_size < 1)
         return CODEC_INSUFFICIENT_BUFFER;
@@ -661,4 +685,3 @@ int64_t decode_uint64 ( uint8_t const* buf_start, uint8_t const* buf_xend, uint6
 
     return ret;
 }
-
